+++ morib/src/main/java/org/morib/server/domain/task/application/ClassifyTaskService.java
@@ -1,11 +1,8 @@
-package org.morib.server.api.homeViewApi.service.classify.task;
+package org.morib.server.domain.task.application;
 
-import org.morib.server.domain.category.infra.Category;
 import org.morib.server.domain.task.infra.Task;
 
 import java.time.LocalDate;
-import java.util.List;
-import java.util.Map;
 
 public interface ClassifyTaskService {
     int classifyTimerByTask(LocalDate date, Task task);
