+++ morib/src/main/java/org/morib/server/domain/task/application/FetchTaskService.java
@@ -1,0 +1,6 @@
+package org.morib.server.domain.task.application;
+
+
+public interface FetchTaskService {
+    void fetch();
+}
