+++ morib/src/main/java/org/morib/server/domain/task/application/FetchTaskServiceImpl.java
@@ -1,4 +1,4 @@
-package org.morib.server.api.homeViewApi.service.fetch.task;
+package org.morib.server.domain.task.application;
 
 import lombok.RequiredArgsConstructor;
 import org.morib.server.domain.task.application.TaskGateway;
